Compile kernel with 4.9
The ROM itself is compiled with 5.1, but when I tried to compile the kernel with 5.1, I literally ran into over 100 different errors. I was able to bypass most of them with cflags, because they were mainly warnings that were being treated as errors. However, I noticed a fairly significant increase in battery drain vs when the kernel is compiled with 4.9. So for now, we're sticking with 4.9. As 5.1 progresses, I will revisit it